Data.Time.JapaneseCalendar.Format

Documentation

formatTempoDate :: String -> TempoDate -> Either String String

a formatter for TempoDate.

formatTempoDate "%y年%M月%d日" converts a TempoDate into a String like "2000年1月2日".

%y

year (e.g. "2000")

%L

leap month sign, "" or "閏"

%m

month, "1" - "12"

%M

same as %L%m (e.g. "閏10")

%b

month in Japanese (e.g. "神無月")

%B

same as %L%b (e.g. "閏神無月")

%d

day, "1" - "31"

padding qualifiers like printf in C are also available.

%2m

pad with two spaces

%02m

pad with two zeros

qFormatTempoDate :: QuasiQuoter

a Template Haskell version of formatTempoDate, which generates a formatter at compile-time.

if a given format is invalid, compilation will be failed.

[qFormatTempoDate|fmt|] is semantically the same as formatTempoDate "fmt".