Data.Time.JapaneseCalendar.Internal.Moon

Documentation

moonEclipticLongitude :: UTCTime -> Double

calculates ecliptic longitude of the moon in degrees from a UTC time.

moonPhase :: UTCTime -> Double

calculates the moon phase in degrees from a UTC time.

nearestNewMoon :: UTCTime -> UTCTime

estimates the nearest day of new moon.

averageMoonPhaseCycle :: Fractional a => a

the average cycle of the moon phase.