Data.Time.JapaneseCalendar.SolarTerm

Documentation

data SolarTerm

the twenty-four solar terms

Constructors

VernalEquinox
春分

ClearAndBright
清明

GrainRain
穀雨

StartOfSummer
立夏

GrainFull
小満

GrainInEar
芒種

SummerSolstice
夏至

MinorHeat
小暑

MajorHeat
大暑

StartOfAutumn
立秋

LimitOfHeat
処暑

WhiteDew
白露

AutumnalEquinox
秋分

ColdDew
寒露

FrostDescent
霜降

StartOfWinter
立冬

MinorSnow
小雪

MajorSnow
大雪

WinterSolstice
冬至

MinorCold
小寒

MajorCold
大寒

StartOfSpring
立春

RainWater
雨水

AwakeningOfInsects
啓蟄

isSegmentPoint :: SolarTerm -> Bool

tests whether a solar term is a segment point, sekki

isCenterPoint :: SolarTerm -> Bool

tests whether a solar term is a center point, chuki

solarTerm :: TimeZone -> Day -> Maybe SolarTerm

returns a solar term of a specified day

nearestSolarTerm :: TimeZone -> Day -> (SolarTerm, Day)

finds the nearest solar term

solarTermsFrom :: TimeZone -> Day -> [(SolarTerm, Day)]

creates a list of solar terms starting with the nearest solar term from a specified day

findNearestSolarTerm :: TimeZone -> SolarTerm -> Day -> Day

finds the nearest specified solar term