Data.Time.JapaneseCalendar.TempoCalendar

Documentation

data TempoMonthType

a type of a month in the Tempo calendar

Constructors

Mutsuki
睦月

Kisaragi
如月

Yayoi
弥生

Uzuki
卯月

Satsuki
皐月

Minazuki
水無月

Fumizuki
文月

Hazuki
葉月

Nagatsuki
長月

Kannazuki
神無月

Shimotsuki
霜月

Shiwasu
師走

data TempoMonth

a month in the Tempo calendar

Constructors

CommonMonth

Fields

tempoMonthType :: TempoMonthType

LeapMonth

Fields

tempoMonthType :: TempoMonthType

data TempoDate

a date in the Tempo calendar

Constructors

TempoDate

Fields

tempoYear :: Integer

tempoMonth :: TempoMonth

tempoDay :: Int

monthNumber :: TempoMonth -> Int

gets the number of the month which starts from one

isLeapMonth :: TempoMonth -> Bool

tests whether a month is a leap month

tempoMonthToJapaneseName :: TempoMonth -> String

returns a Japanese name of a month in the Tempo calendar

tempoMonthFromJapaneseName :: String -> Maybe TempoMonth

converts a Japanese name of a month in the Tempo calendar into a TempoMonth

nextTempoMonthType :: TempoMonthType -> TempoMonthType

the cyclic successor of TempoMonthType

previousTempoMonthType :: TempoMonthType -> TempoMonthType

the cyclic predecessor of TempoMonthType

tempoDate :: TimeZone -> Day -> Maybe TempoDate

converts a date into a date in the Tempo calendar