Data.Time.JapaneseCalendar.Internal.AstroUtils

Documentation

solveSawtooth

Arguments

:: Double
the amplitude of f

-> Double
the approximated cycle of f

-> (Double -> Double)
the function f, which has a similar form to sawtooth

-> Int
the max iteration number

-> Double
the tolerance of x

-> Double
the initial guess of x

-> Double
the right hand side value y of the equation

-> Double
the solution of the equation

solves an equation f x = y, in case f has a similar form to sawtooth.